[CONVERT MARKDOWN] updated script to convert HBS to MD using Turndown (partial) + removed conversion script based on node-html-markdown/NHM because it was messing up with spacing, so better to focus on a single tool
