+++ website-html-to-markdown/package.json
@@ -11,7 +11,7 @@
   "license": "MPL-2.0",
   "author": "HashiCorp Design Systems <design-systems@hashicorp.com>",
   "scripts": {
-    "generate": "yarn prepare-source-folder && yarn copy-source-files && yarn identify-sections && yarn split-files && yarn replace-linkto-with-a && yarn refactor-codeblocks && yarn preprocess-files && yarn rename-files-as-markdown && yarn prettify-files && yarn prepare-dest-folder && yarn copy-markdown-files && yarn copy-assets-folder && yarn copy-extra-source-files",
+    "generate": "yarn prepare-source-folder && yarn copy-source-files && yarn identify-sections && yarn split-files && yarn replace-linkto-with-a && yarn refactor-codeblocks && yarn preprocess-files && yarn convert-to-markdown && yarn prettify-files && yarn prepare-dest-folder && yarn copy-markdown-files && yarn copy-assets-folder && yarn copy-extra-source-files",
     "prepare-source-folder": "rm -fr source && mkdir source && mkdir source/original",
     "copy-source-files": "cp -r ../packages/components/tests/dummy/app/templates/* source/original",
     "identify-sections": "echo '\n==========\nIdentifying <section> blocks...\n' && node codemods/bin/cli.js identify-sections ./source/original/**/*.hbs",
@@ -19,14 +19,12 @@
     "replace-linkto-with-a": "echo '\n==========\nReplacing <LinkTo> with <a> links...\n' && node codemods/bin/cli.js replace-linkto-with-a ./source/split-files/**/*.hbs",
     "refactor-codeblocks": "echo '\n==========\nRefactoring <CodeBlock> instances...\n' && node codemods/bin/cli.js refactor-codeblocks ./source/split-files/**/03--how-to-use.hbs",
     "preprocess-files": "ts-node --transpile-only ./scripts/preprocess-files",
-    "rename-files-as-markdown": "ts-node --transpile-only ./scripts/rename-files-as-markdown",
+    "convert-to-markdown": "ts-node --transpile-only ./scripts/convert-to-markdown",
     "prettify-files": "echo '\n==========\nPrettifying markdown files...\n' && npx prettier --loglevel warn --write source/split-files",
     "prepare-dest-folder": "rm -fr ../website/docs/components && mkdir ../website/docs/components && rm -fr ../website/docs/content && mkdir ../website/docs/content && rm -fr ../website/docs/foundations && mkdir ../website/docs/foundations && rm -fr ../website/docs/overrides && mkdir ../website/docs/overrides&& rm -fr ../website/docs/utilities && mkdir ../website/docs/utilities",
-    "copy-markdown-files": "echo '\n==========\nCopying markdown files to `docs` folder...\n' && cp -r ./source/split-files/* ../website/docs/",
+    "copy-markdown-files": "echo '\n==========\nCopying markdown files to `docs` folder...\n' && cp -r ./source/markdown/* ../website/docs/",
     "copy-assets-folder": "rm -fr ../website/public/assets/images && cp -r ../packages/components/tests/dummy/public/assets/images ../website/public/assets/",
-    "copy-extra-source-files": "cp -r ./moveover/* ../website/docs/",
-    "convert:turndown": "ts-node --transpile-only ./turndown/convert",
-    "convert:nhm": "ts-node --transpile-only ./nhm/convert"
+    "copy-extra-source-files": "cp -r ./moveover/* ../website/docs/"
   },
   "devDependencies": {
     "@types/fs-extra": "^9.0.13",
